+++ src/main/java/roomescape/service/dto/ReservationTimeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.web.dto;
+package roomescape.service.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonFormat.Shape;
