+++ src/main/java/roomescape/service/dto/ThemeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.web.dto;
+package roomescape.service.dto;
 
 import roomescape.domain.Theme;
